features/presentation.py
import os
try:
    from pptx import Presentation
    from pdf2image import convert_from_path # Alternative for PPT if needed
except ImportError:
    pass
import cv2
import numpy as np

class PresentationService:
    def __init__(self):
        self.slides = []
        self.current_slide_index = 0
        self.mode = "NONE" # "PPT" or "WEB"

    def load_ppt(self, ppt_path):
        """
        Loads a PPT file. 
        Note: Real-time PPT rendering in OpenCV/Qt can be complex. 
        A common cheat is to convert slides to images.
        """
        # Placeholder for pptx to image conversion
        # In a real app, we might use COM dispatch on Windows or a library like pdf2image
        # For this demo, we'll look for an 'images' folder or similar
        pass

    def next_slide(self):
        if self.slides:
            self.current_slide_index = (self.current_slide_index + 1) % len(self.slides)

    def prev_slide(self):
        if self.slides:
            self.current_slide_index = (self.current_slide_index - 1) % len(self.slides)

    def get_current_slide(self, width, height):
        if self.slides:
            slide = self.slides[self.current_slide_index]
            return cv2.resize(slide, (width, height))
        return None
